refactor(app-headless-cms-scheduler): rename decorator file to match its export + document conventions
- Rename ContentEntriesPresenterDecorator.ts -> ContentEntriesPresenterSchedulingDecorator.ts
  so the filename matches the exported symbol.
- AGENTS.md: filename must match an exported symbol; presenter hooks use the
  Presenter suffix and are resolved via a dedicated hook.
